+++ src/templates/welcome-email/index.html
@@ -177,7 +177,7 @@ table, td { color: #000000; } #u_body a { color: #0000ee; text-decoration: under
       <td class="v-container-padding-padding" style="overflow-wrap:break-word;word-break:break-word;padding:40px 30px;font-family:'Open Sans',sans-serif;" align="left">
         
   <div class="v-line-height v-font-size" style="line-height: 180%; text-align: justify; word-wrap: break-word;">
-    <p style="line-height: 180%; font-size: 14px; text-align: center;">Ya fue aprobada tu compra [id-transaction] esperamos sigas aprendiendo mas con nosotros</p>
+    <p style="line-height: 180%; font-size: 14px; text-align: center;">Ya fue aprobada tu compra [id-order] esperamos sigas aprendiendo mas con nosotros</p>
   </div>
 
       </td>
